Address copilot review comments, remove unused imports
