+++ tap_activecampaign/transform.py
@@ -1,4 +1,3 @@
-import re
 import humps
 import singer
